pyproject.toml
[project]
name = "groundtruth-census-tools"
version = "0.1.0"
description = "Coordinate-to-tract Census Reporter lookup utilities"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "uvicorn>=0.35.0,<1.0.0",
  "pydantic>=2.0.0",
  "google-generativeai>=0.8.0",
  "requests>=2.32.0",
  "python-dotenv>=1.0.0,<2.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[dependency-groups]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "pytest-asyncio>=0.23.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
]
